thresholdr: Estimate Optimal Probability Classification Thresholds for Diagnostic Models

Diagnostic classification models are confirmatory latent class models that provide probabilities of proficiency for respondents on each of the assessed skills (Bradshaw, 2016 doi:10.1002/9781118956588.ch13 ). Often, results based on these models are reported as classifications, using 0.5 as the probability threshold for classification (Bradshaw & Levy, 2019 doi:10.1111/emip.12247 ). Instead, we can use resampling methods to estimate a different classification threshold that optimizes sensitivity and specificity.

See also

Author

Maintainer: W. Jake Thompson wjakethompson@gmail.com (ORCID)

Authors:

W. Jake Thompson wjakethompson@gmail.com (ORCID)

Jonathan A. Pedroza

Other contributors:

Andrew Heiss (shapes_to_muphi() and muphi_to_shapes()) [copyright holder]

University of Kansas [copyright holder]